chore: apply comments
