+++ agent/monitor.go
@@ -16,5 +16,5 @@ type PluginInfo struct {
 type Monitor interface {
 	RecordRun(ctx context.Context, run Run)
 	RecordPlugin(ctx context.Context, pluginInfo PluginInfo)
-	RecordPluginRetryCount(ctx context.Context, pluginInfo PluginInfo)
+	RecordSinkRetryCount(ctx context.Context, pluginInfo PluginInfo)
 }
